third

+++ Domain/Dtos/StudentCountDto.cs
@@ -5,5 +5,5 @@ namespace Domain.Dtos;
 public class StudentCountDto
 {
     public string Title { get; set; }
-    public int Count { get; set; }
+    public double Count { get; set; }
 }
